Improved authorization in patientController

+++ app/views/patient/profile.php
@@ -6,12 +6,18 @@ $this->start('body');
 <div class="container">
 
     <div>
-        <?php if (isset($_SESSION['msg']) || $msg != null)
-        {
-            echo $_SESSION['msg'];
-            unset($_SESSION['msg']);
-        }elseif ($msg != null){
-            echo $msg;
+        <?php if (isset($_SESSION['msg']) || $msg != "" || isset($_REQUEST['msg'])){
+            if (isset($_SESSION['msg']))
+            {
+                echo $_SESSION['msg'];
+                unset($_SESSION['msg']);
+            }elseif ($msg != null){
+                echo "<div>".$msg."</div>";
+            }
+            elseif (isset($_REQUEST['msg'])){
+                echo $_REQUEST['msg'];
+                unset($_REQUEST['msg']);
+            }
         }
         ?>
     </div>
@@ -56,8 +62,11 @@ $this->start('body');
             echo "<td>".Role::all()->where('id', $patient->role_id)->first()->roleType."</td>";
 
             echo "<td><a href='".SROOT."patient/edit/" . $patient->id. "' title='Edit Record' class='btn btn-warning btn-xs' data-toggle='tooltip'><i class='fa fa-edit'></i></a></td>";
-            echo "<td><a href='".SROOT."patient/profile/" . $patient->id. "' title='Go to Profile' class='btn btn-info btn-xs' data-toggle='tooltip'><i class='fa fa-user'></i></a></td>";
-            echo "<td><a href='".SROOT."patient/delete/" . $patient->id. "' title='Delete record' class='btn btn-danger btn-xs' data-toggle='tooltip'><i class='fa fa-trash'></i></a></td>";
+            //echo "<td><a href='".SROOT."patient/profile/" . $patient->id. "' title='Go to Profile' class='btn btn-info btn-xs' data-toggle='tooltip'><i class='fa fa-user'></i></a></td>";
+            if (LoginHelper::isAdmin()){
+                echo "<td><a href='".SROOT."patient/delete/" . $patient->id. "' title='Delete record' class='btn btn-danger btn-xs' data-toggle='tooltip'><i class='fa fa-trash'></i></a></td>";
+            }
+
             echo "</tr>";
 
         ?>
